+++ script/bcar.js
@@ -1,4 +1,4 @@
-const BCAR_Version = '0.7.3';
+const BCAR_Version = '0.7.4';
 const BCAR_Settings_Version = 8;
 
 const ICONS = Object.freeze({
@@ -45,15 +45,25 @@ var bcModSDK=function(){"use strict";const e="1.1.0";function o(e){alert("Mod ER
   await waitFor(() => ServerIsConnected && ServerSocket);
   //end of do not touch
   const bcarSettingsKey = () => `bcarSettings.${Player?.AccountName}`;
-    const subcommands = ["animationbuttons", "animal", "animalhelp", "arousal", "arousalhelp", "changelog", "cat", "delete1", "delete2", "delete3", "dog", "ear1", "ear2", "eardelete", "eardelay", "earhelp", "earwiggle", "earwigglecount", "emotehelp", "emoteear", "emotetail", "expressionhelp", "expression", "fox", "female", "help", "load1", "load2", "load3", "lowerleft", "lowerright", "male", "misc", "mouse", "other", "profile1", "profile2", "profile3", "profilehelp", "reset", "save1", "save2", "save3", "settings", "status", "tail1", "tail2", "tailhelp", "taildelay", "taildelete", "tailwag", "tailwagcount", "timerhelp", "timer", "upperleft","versions", "wing1", "wing2", "wingdelay", "wingdelete", "wingflapcount", "winghelp", "wingflap"];
+    const subcommands = ["animationbuttons", "animal", "animalhelp", "arousal", "arousalhelp", "changelog", "cat", "delete1", "delete2", "delete3", "dog", "ear1", "ear2", "eardelete", "eardelay", "earhelp", "earwiggle", "earwigglecount", "emotehelp", "emoteear", "emotetail", "expressionhelp", "expression", "fox", "female", "fly", "help", "human", "land", "load1", "load2", "load3", "lowerleft", "lowerright", "male", "misc", "mouse", "other", "profile1", "profile2", "profile3", "profilehelp", "reset", "save1", "save2", "save3", "settings", "status", "tail1", "tail2", "tailhelp", "taildelay", "taildelete", "tailwag", "tailwagcount", "timerhelp", "timer", "upperleft","versions", "wing1", "wing2", "wingdelay", "wingdelete", "wingflapcount", "winghelp", "wingflap"];
     const w = window;
     const BCAR_CHANGELOG =
           "BCAR+ v" + BCAR_Version +
+          "<br>- Added the ''animal'' human" +
           "<br>- Settings page (wip)" +
-          "<br>  - The settings page will now be exited properly" +
+          "<br>  - Added option to change gender" +
+          "<br>  - Added option to change animal type" +
+          "<br>  - Added option to change the animation button position" +
+          "<br>- Added 'flutter' as a verb as well as 'flap' and use it for fairy, pixie and bee wings by <a href='https://github.com/elliethepink' target='_blank'>@elliethepink</a>" +
+          "<br>- Added commands to start and stop flying <a href='https://github.com/elliethepink' target='_blank'>@elliethepink</a>" +
+          "<br>- Added more items that prevent flying in addition to neck restraints, factor out the check for whether the player can fly into a separate function & also check the player has wings by <a href='https://github.com/elliethepink' target='_blank'>@elliethepink</a>" +
+          "<br>- Fixed a crash if you tried to fly but had never set an emoticon by <a href='https://github.com/elliethepink' target='_blank'>@elliethepink</a>" +
+          "<br>- Fixed that AccountUpdate is being sent, reduces the message load caused by BCAR+ by <a href='https://github.com/elliethepink' target='_blank'>@elliethepink</a>" +
+          "<br>- Cleaned up the code" +
           "<br>" +
-          "<br>BCAR+ v0.7.2(a):" +
-          "<br>- BCAR+ didn't load at all"
+          "<br>BCAR+ v0.7.3" +
+          "<br>- Settings page (wip)" +
+          "<br>The settings page will now be exited properly"
 
 
     function copy_object(o) {
@@ -84,6 +94,8 @@ var bcModSDK=function(){"use strict";const e="1.1.0";function o(e){alert("Mod ER
         );
         Player.FocusGroup = focusGroup;
     }
+
+    const getWingVerb = () => ["FairyWings", "BeeWings", "PixieWings"].includes(InventoryGet(Player,"Wings")?.Asset?.Name) ? "flutters" : "flaps";
 
     function BCARChatRoomClick() {
         modApi.hookFunction('ChatRoomClick', 4, (args, next) => {
@@ -144,7 +156,7 @@ var bcModSDK=function(){"use strict";const e="1.1.0";function o(e){alert("Mod ER
                                 { Tag: "Beep", Text: "msg" },
                                 { Tag: "Biep", Text: "msg" },
                                 { Tag: "Sonner", Text: "msg" },
-                                { Tag: "msg", Text: CharacterNickname(Player) + " flaps " + Player.BCAR.bcarSettings.genderDefault.capPossessive.toLocaleLowerCase() + " wings." }
+                                { Tag: "msg", Text: CharacterNickname(Player) + " " + getWingVerb() + " " + Player.BCAR.bcarSettings.genderDefault.capPossessive.toLocaleLowerCase() + " wings." }
                             ]
                         });
                         WingFlap();
@@ -898,7 +910,8 @@ const TriggerAdditions = [
           for(let i=0; i < numberWiggles; i++)
           {
               setTimeout(function() {
-                  InventoryWear(Player, earsVariations[i%earsVariations.length], "HairAccessory2", earsColor[i%earsColor.length]);
+                  InventoryWear(Player, earsVariations[i%earsVariations.length], "HairAccessory2", earsColor[i%earsColor.length], undefined, undefined, undefined, false);
+                  CharacterRefresh(Player, false);
                   ChatRoomCharacterItemUpdate(Player, "HairAccessory2");
               }, i * delay);
           }
@@ -914,7 +927,8 @@ const TriggerAdditions = [
             for(let i=0; i < numberWags; i++)
             {
                 setTimeout(function() {
-                    InventoryWear(Player, tailsVariations[i%tailsVariations.length], "TailStraps", tailsColor[i%tailsColor.length]);
+                    InventoryWear(Player, tailsVariations[i%tailsVariations.length], "TailStraps", tailsColor[i%tailsColor.length], undefined, undefined, undefined, false);
+                    CharacterRefresh(Player, false);
                     ChatRoomCharacterItemUpdate(Player, "TailStraps");
                 }, i * delay);
             }
@@ -930,7 +944,8 @@ const TriggerAdditions = [
             for(let i=0; i < numberFlaps; i++)
             {
                 setTimeout(function() {
-                    InventoryWear(Player, wingsVariations[i%wingsVariations.length], "Wings", wingsColor[i%wingsColor.length]);
+                    InventoryWear(Player, wingsVariations[i%wingsVariations.length], "Wings", wingsColor[i%wingsColor.length], undefined, undefined, undefined, false);
+                    CharacterRefresh(Player, false);
                     ChatRoomCharacterItemUpdate(Player, "Wings");
                 }, i * delay);
             }
@@ -944,10 +959,52 @@ const TriggerAdditions = [
         }
     }
 
+    function TryFly() {
+        if (!InventoryGet(Player,"Wings")) {
+            ChatRoomSendLocal(
+                `<p style='background-color:#000452;color:#EEEEEE;'><b>Bondage Club Auto React +</b>
+                <br>You need wings to fly!</p>`.replaceAll('\n', ''), wt.info
+            );
+            ServerSend("ChatRoomChat", { Content: "Beep", Type: "Action", Dictionary: [{Tag: "Beep", Text: `${CharacterNickname(Player)} tried to fly without wings.` }]});
+            return false;
+        }
+
+        if (Player.BCAR.bcarSettings.wingFlappingEnable !== true){
+            ChatRoomSendLocal(
+                `<p style='background-color:#000452;color:#EEEEEE;'><b>Bondage Club Auto React +</b>
+                <br>Wing flaping must be enabled to fly.
+                </p>`.replaceAll('\n', ''), wt.info
+            );
+            return false;
+        }
+
+        // Many items don't tether/chain/etc but ought to prevent flying
+        const otherPreventingItemNames = ["CeilingShackles", "FloorShackles", "SuspensionCuffs", "CeilingRope", "CeilingChain", "BallChain"];
+        const preventingItem = Player.Appearance.find(i => (
+            ["Tethered", "Chained", "Mounted", "Enclose", "IsLeashed", "IsChained"].some(e => InventoryItemHasEffect(i, e))
+            || otherPreventingItemNames.includes(i.Asset.Name)
+        ));
+        if (preventingItem) {
+          ChatRoomSendLocal(
+              `<p style='background-color:#000452;color:#EEEEEE;'><b>Bondage Club Auto React +</b>
+              <br>You can't fly because the ${preventingItem.Asset.Description} holds you down.
+              </p>`.replaceAll('\n', ''), wt.info
+          )
+          ServerSend("ChatRoomChat", { Content: "Beep", Type: "Action", Dictionary: [{Tag: "Beep", Text:  `${CharacterNickname(Player)} tried to fly while ${Player.BCAR.bcarSettings.genderDefault.capPronoun.toLocaleLowerCase()} is held down by a ${preventingItem.Asset.Description}.` }]});
+            return false;
+        } else {
+            Fly();
+            WingFlap();
+            return true;
+        }
+    }
+
     function Fly(){
         if(Player.BCAR.bcarSettings.wingFlappingEnable === true){
             CharacterSetActivePose(Player, "LegsClosed");
-            InventoryGet(Player, 'Emoticon').Property.OverrideHeight = { Height: +70 };
+            const emoticon = InventoryGet(Player, 'Emoticon');
+            if (emoticon.Property === undefined) emoticon.Property = {};
+            emoticon.Property.OverrideHeight = { Height: +70 };
         }
     }
 
@@ -1069,7 +1126,7 @@ const TriggerAdditions = [
 
     if(data.Type === "Emote" && data.Sender === Player.MemberNumber){
           var flapMessage = data.Content;
-          let patterns = [/flaps.*wings/mi, /wings.*flapping/mi, /flapping.*wings/mi, /wings.*flap/mi] ; // matches {<any> flaps <any> wings <any>}
+          let patterns = [/(flaps|flutters).*wings/mi, /wings.*(flapping|fluttering)/mi, /(flapping|fluttering).*wings/mi, /wings.*(flap|flutter)/mi] ; // matches {<any> flaps/flutter <any> wings <any>}
           let result = patterns.find(pattern => pattern.test(flapMessage));
           if(result){
               WingFlap();
@@ -1080,21 +1137,9 @@ const TriggerAdditions = [
           var flyMessage = data.Content;
           let patterns = [/begins.*fly/mi, /starts.*flying/mi] ; // matches {<any> begins <any> fly <any>}
           let result = patterns.find(pattern => pattern.test(flyMessage));
-          let NeckRestraints = InventoryGet(Player, "ItemNeckRestraints");
+
           if(result){
-              if (InventoryGet(Player, "ItemNeckRestraints")){
-                  //console.log("IF")
-                  ChatRoomSendLocal(
-                    `<p style='background-color:#000452;color:#EEEEEE;'><b>Bondage Club Auto React +</b>
-                    "You can't fly because the ${NeckRestraints.Asset.Description} holds you down.</p>`, wt.info
-                )
-                ServerSend("ChatRoomChat", { Content: "Beep", Type: "Action", Dictionary: [{Tag: "Beep", Text: `${CharacterNickname(Player)} tried to fly while ${Player.BCAR.bcarSettings.genderDefault.capPronoun.toLocaleLowerCase()} is held down by a ${NeckRestraints.Asset.Description}.` }]});
-            }
-              else {
-                  //console.log("ELSE")
-                  Fly();
-                  WingFlap();
-              }
+              TryFly();
           }
       }
 
@@ -1935,6 +1980,42 @@ function CommandWingDelete(argsList)
         }
     }
 
+function CommandFly(argsList)
+    {
+        const cmd = argsList[0];
+        switch (cmd) {
+            case 'fly':
+                if (TryFly()) {
+                    ServerSend("ChatRoomChat", {
+                        Content: "Beep",
+                        Type: "Action",
+                        Target: null,
+                        Dictionary: [
+                            { Tag: "Beep", Text: "msg" },
+                            { Tag: "Biep", Text: "msg" },
+                            { Tag: "Sonner", Text: "msg" },
+                            { Tag: "msg", Text: `${CharacterNickname(Player)} starts flying.` }
+                        ]
+                    });
+                }
+                break;
+            case 'land':
+                ServerSend("ChatRoomChat", {
+                    Content: "Beep",
+                    Type: "Action",
+                    Target: null,
+                    Dictionary: [
+                        { Tag: "Beep", Text: "msg" },
+                        { Tag: "Biep", Text: "msg" },
+                        { Tag: "Sonner", Text: "msg" },
+                        { Tag: "msg", Text: `${CharacterNickname(Player)} lands back on the ground.` }
+                    ]
+                });
+                Landing();
+                break;
+        }
+}
+
 function CommandWingHelp(argsList)
     {
         const cmd = argsList[0];
@@ -1943,9 +2024,9 @@ function CommandWingHelp(argsList)
                 ChatRoomSendLocal(
                 `<p style='background-color:#000452;color:#EEEEEE;'><b>Bondage Club Auto React +</b>
                 <br>Wing instructions:
-                <br>First equip the main wings you want to wear primarily in the "Wings" slot in your wardrobe. Type "/bcar wing1" in the chat to save the main wings.
-                <br>For your wings to wiggle follow the same steps and equip a different type of wings to use as your secondary. Type "/bcar wing2" in the chat to save the secondary wings.
-                <br>To let your wings flap type an emote anything that includes the words "flaps" and "wings".
+                <br>First equip the main wings you want to wear primarily in the ''Wings'' slot in your wardrobe. Type ''/bcar wing1'' in the chat to save the main wings.
+                <br>For your wings to wiggle follow the same steps and equip a different type of wings to use as your secondary. Type ''/bcar wing2'' in the chat to save the secondary wings.
+                <br>To let your wings flap type an emote anything that includes one of the words ''flaps'' or ''flutters'' and the word ''wings''.
                 <br>
                 <br>Commands:
                 <br>/bcar wing1 - Saves the primary wings.
@@ -1954,12 +2035,16 @@ function CommandWingHelp(argsList)
                 <br>/bcar wingflapcount - Determines the number of flaps.
                 <br>/bcar wingdelay - Determines the flapping speed.
                 <br>/bcar wingdelete - Removes the wings.
+                <br>/bcar fly - Starts flying
+                <br>/bcar land - Stops flying
                 <br>
                 <br>Examples:
                 <br><i>*flaps her wings
+                <br>*flutters her wings
                 <br>*is flapping her wings
                 <br>*lets her wings flap
-                <br>*spreads her wings, flapping with them</i></p>`.replaceAll('\n', ''), wt.help
+                <br>*spreads her wings, flapping with them</i>
+                </p>`.replaceAll('\n', ''), wt.help
                 );
         }
     }
@@ -2092,7 +2177,7 @@ function CommandProfile(argsList)
 //End of Profile Commands
 
 //Misc Commands
-function CommandAnimationButtons(argsList)
+function CommandAnimationButtons(argsList, showMessage = true)
     {
         const cmd = argsList[0].toLocaleLowerCase();;
         const s = Player?.BCAR?.bcarSettings;
@@ -2115,7 +2200,7 @@ function CommandAnimationButtons(argsList)
                     `<p style='background-color:#630A0A;color:#EEEEEE;'><b>Bondage Club Auto React +</b>
                     <br>Animation Buttons are now disabled!
                     </p>`.replaceAll('\n', ''), wt.info
-                    );
+                        );
                     BCARChatRoomMenuDraw();
                     BCARChatRoomClick();
                     bcarSettingsSave();
@@ -2123,28 +2208,32 @@ function CommandAnimationButtons(argsList)
                 break;
             case 'upperleft' : case 'lowerleft' : case 'lowerright' :
                 Player.BCAR.bcarSettings.animationButtonsPosition = cmd
-                ChatRoomSendLocal(
-                `<p style='background-color:#000452;color:#EEEEEE;'><b>Bondage Club Auto React +</b>
-                <br>Animation Buttons moved to ${cmd}!
-                </p>`.replaceAll('\n', ''), wt.info
-                );
+                if (showMessage){
+                    ChatRoomSendLocal(
+                    `<p style='background-color:#000452;color:#EEEEEE;'><b>Bondage Club Auto React +</b>
+                    <br>Animation Buttons moved to ${cmd}!
+                    </p>`.replaceAll('\n', ''), wt.info
+                    );
+                }
                 BCARChatRoomMenuDraw();
                 BCARChatRoomClick();
                 bcarSettingsSave();
         }
     }
 
-function CommandAnimals(argsList)
+function CommandAnimals(argsList, showMessage = true)
     {
         const cmd = argsList[0].toLocaleLowerCase();
         switch (cmd) {
-            case 'cat': case 'dog': case 'fox': case 'mouse':
+            case 'cat': case 'dog': case 'fox': case 'human': case 'mouse':
                 Player.BCAR.bcarSettings.animal = cmd
-                ChatRoomSendLocal(
-                `<p style='background-color:#000452;color:#EEEEEE;'><b>Bondage Club Auto React +</b>
-                <br><span style="text-transform:capitalize">${cmd}</span> reactions are selected!
-                </p>`.replaceAll('\n', ''), wt.info
-                );
+                if (showMessage){
+                    ChatRoomSendLocal(
+                    `<p style='background-color:#000452;color:#EEEEEE;'><b>Bondage Club Auto React +</b>
+                    <br><span style="text-transform:capitalize">${cmd}</span> reactions are selected!
+                    </p>`.replaceAll('\n', ''), wt.info
+                    );
+                }
                 bcarSettingsSave();
                 break;
             case 'animal': case 'animalhelp' :
@@ -2491,43 +2580,91 @@ function CommandExpression(argsList)
         }
     }
 
-function CommandGenderToggle(argsList)
+function CommandGenderToggle(argsList, showMessage = true)
     {
-        let toggle = argsList[0];
+        const cmd = argsList[0];
+        const s = Player?.BCAR?.bcarSettings;
+        switch (cmd) {
+                case 'male':
+                s.genderDefault.gender = "Male";
+                s.genderDefault.capPronoun = "He";
+                s.genderDefault.capIntensive = "Him";
+                s.genderDefault.capPossessive = "His";
+                if (showMessage){
+                    ChatRoomSendLocal(
+                    `<p style='background-color:#000452;color:#EEEEEE;'><b>Bondage Club Auto React +</b>
+                    <br>The reactions refer to ${CharacterNickname(Player)} as "he" now!
+                    </p>`.replaceAll('\n', ''), wt.info
+                    );
+                }
+                break;
+            case 'female':
+                s.genderDefault.gender = "Female";
+                s.genderDefault.capPronoun = "She";
+                s.genderDefault.capIntensive = "Her";
+                s.genderDefault.capPossessive = "Her";
+                if (showMessage){
+                    ChatRoomSendLocal(
+                    `<p style='background-color:#000452;color:#EEEEEE;'><b>Bondage Club Auto React +</b>
+                    <br>The reactions refer to ${CharacterNickname(Player)} as "she" now!
+                    </p>`.replaceAll('\n', ''), wt.info
+                    );
+                }
+                break;
+            case 'other':
+                s.genderDefault.gender = "Non-Binary";
+                s.genderDefault.capPronoun = "They";
+                s.genderDefault.capIntensive = "Them";
+                s.genderDefault.capPossessive = "Their";
+                if (showMessage){
+                    ChatRoomSendLocal(
+                    `<p style='background-color:#000452;color:#EEEEEE;'><b>Bondage Club Auto React +</b>
+                    <br>The reactions refer to ${CharacterNickname(Player)} as "they" now!
+                    </p>`.replaceAll('\n', ''), wt.info
+                    );
+                }
+        }
+        /*let toggle = argsList[0];
 
         if (toggle === "male") {
             Player.BCAR.bcarSettings.genderDefault.gender = "Male";
             Player.BCAR.bcarSettings.genderDefault.capPronoun = "He";
             Player.BCAR.bcarSettings.genderDefault.capIntensive = "Him";
             Player.BCAR.bcarSettings.genderDefault.capPossessive = "His";
-            ChatRoomSendLocal(
+            if (showMessage){
+                ChatRoomSendLocal(
                 `<p style='background-color:#000452;color:#EEEEEE;'><b>Bondage Club Auto React +</b>
                 <br>The reactions refer to ${CharacterNickname(Player)} as "he" now!
                 </p>`.replaceAll('\n', ''), wt.info
-            );
+                );
+            }
         }
         else if (toggle === "female") {
             Player.BCAR.bcarSettings.genderDefault.gender = "Female";
             Player.BCAR.bcarSettings.genderDefault.capPronoun = "She";
             Player.BCAR.bcarSettings.genderDefault.capIntensive = "Her";
             Player.BCAR.bcarSettings.genderDefault.capPossessive = "Her";
-            ChatRoomSendLocal(
+            if (showMessage){
+                ChatRoomSendLocal(
                 `<p style='background-color:#000452;color:#EEEEEE;'><b>Bondage Club Auto React +</b>
                 <br>The reactions refer to ${CharacterNickname(Player)} as "she" now!
                 </p>`.replaceAll('\n', ''), wt.info
-            );
+                );
+            }
         }
         else if (toggle === "other") {
             Player.BCAR.bcarSettings.genderDefault.gender = "Non-Binary";
             Player.BCAR.bcarSettings.genderDefault.capPronoun = "They";
             Player.BCAR.bcarSettings.genderDefault.capIntensive = "Them";
             Player.BCAR.bcarSettings.genderDefault.capPossessive = "Their";
-            ChatRoomSendLocal(
+            if (showMessage){
+                ChatRoomSendLocal(
                 `<p style='background-color:#000452;color:#EEEEEE;'><b>Bondage Club Auto React +</b>
                 <br>The reactions refer to ${CharacterNickname(Player)} as "they" now!
                 </p>`.replaceAll('\n', ''), wt.info
-            );
-        }
+                );
+            }
+        } */
         bcarSettingsSave();
     }
 
@@ -2753,6 +2890,7 @@ function CommandVersions(argsList)
                 CommandWingFlapCountChange(args.split(" "));
                 CommandWingFlapDelayChange(args.split(" "));
                 CommandWingDelete(args.split(" "));
+                CommandFly(args.split(" "));
                 CommandWingHelp(args.split(" "));
                 //Profile Commands
                 CommandProfile(args.split(" "));
@@ -2872,8 +3010,132 @@ CommandCombine([
     }
 
     // MAIN MENU
+    const settings = {
+        gender: {
+            options: [
+                "male",
+                "female",
+                "other",
+            ],
+            labels: [
+                "Male",
+                "Female",
+                "Non-Binary",
+            ],
+            selected: 0,
+            x: 950,
+            y: getYPos(0),
+            w: 400,
+            h: 85,
+        },
+        animal: {
+            options: [
+                "cat",
+                "dog",
+                "fox",
+                "human",
+                "mouse",
+            ],
+            labels: [
+                "Cat",
+                "Dog",
+                "Fox",
+                "Human",
+                "Mouse",
+            ],
+            selected: 0,
+            x: 950,
+            y: getYPos(1),
+            w: 400,
+            h: 85,
+        },
+        animationButtonsPosition: {
+            options: [
+                "upperleft",
+                "lowerleft",
+                "lowerright",
+            ],
+            labels: [
+                "Upper Left",
+                "Lower Left",
+                "Lower Right",
+            ],
+            selected: 0,
+            x: 950,
+            y: getYPos(2),
+            w: 400,
+            h: 85,
+        },
+
+            init: function(name, value) {
+      let setting = this[name];
+      setting.selected = setting.options.indexOf(value);
+      if (setting.selected < 0) setting.selected = setting.labels.indexOf(value);
+      if (setting.selected < 0) setting.selected = 0;
+    },
+    value: function(name) {
+      let setting = this[name];
+      return setting.options[setting.selected];
+    },
+    label: function(name) {
+      let setting = this[name];
+      return setting.labels[setting.selected];
+    },
+    index_next: function(name) {
+      let setting = this[name];
+      let index = setting.selected;
+      index += 1;
+      if (index >= setting.options.length) index = 0;
+      return index;
+    },
+    index_prev: function(name) {
+      let setting = this[name];
+      let index = setting.selected;
+      index -= 1;
+      if (index < 0) index = setting.options.length - 1;
+      return index;
+    },
+    select_next: function(name) {
+      this[name].selected = this.index_next(name);
+    },
+    select_prev: function(name) {
+      this[name].selected = this.index_prev(name);
+    },
+    peek_prev_label: function(name) {
+      return this[name].labels[this.index_prev(name)];
+    },
+    peek_next_label: function(name) {
+      return this[name].labels[this.index_next(name)];
+    },
+    draw: function(name) {
+      let setting = this[name];
+      const save_align = MainCanvas.textAlign;
+      MainCanvas.textAlign = "center";
+      DrawBackNextButton(setting.x, setting.y, setting.w, setting.h, this.label(name), "White", "",
+                               () => this.peek_prev_label(name),
+                               () => this.peek_next_label(name),
+                              );
+        MainCanvas.textAlign = save_align;
+        },
+        click: function(name) {
+            let setting = this[name];
+            if (MouseIn(setting.x, setting.y, setting.w, setting.h)) {
+                if (MouseIn(setting.x, setting.y, (setting.w / 2), setting.h)) {
+                    this.select_prev(name);
+                } else {
+                    this.select_next(name);
+                }
+                return true; // there was a click on this button
+            }
+            return false; // there were no clicks on this button
+        },
+    };
+
     w.PreferenceSubscreenBCARSettingsLoad = function() {
         console.debug("BCAR+ Settings load");
+        settings.init("gender", Player.BCAR?.bcarSettings?.genderDefault?.gender);
+        settings.init("animal", Player.BCAR?.bcarSettings?.animal);
+        settings.init("animationButtonsPosition", Player.BCAR?.bcarSettings?.animationButtonsPosition);
 
     };
 
@@ -2906,7 +3168,8 @@ CommandCombine([
         );
 
         DrawButton(1815, 75, 90, 90, "", "White", "Icons/Exit.png");
-        DrawButton(1540, 770, 400, 85, "", "White", "Icons/Changelog.png", "Open BCAR+ Changelog on GitHub.", true);
+        DrawButton(1540, 770, 400, 85, "", "White", "", "Open BCAR+ Changelog on GitHub.");
+        DrawImageResize("Icons/Changelog.png", 1550, 780, 60, 60);
         DrawTextFit(
             "Open Changelog",
             1635,
@@ -2914,7 +3177,8 @@ CommandCombine([
             380,
             "Black"
         );
-        DrawButton(1540, 870, 400, 85, "", "White", undefined, "Open BCAR+ Wiki on GitHub.", true);
+        DrawButton(1540, 870, 400, 85, "", "White", "", "Open BCAR+ Wiki on GitHub.");
+        DrawImageResize("Icons/Introduction.png", 1550, 880, 60, 60);
         DrawTextFit(
             "Open Wiki",
             1650,
@@ -2931,6 +3195,7 @@ CommandCombine([
             "Black"
         );
         DrawButton(500, getYPos(1), 400, 85, "", "White"/*, "Assets/Female3DCG/HairAccessory1/Preview/KittenEars1.png"*/);
+        //DrawImageResize("Assets/Female3DCG/HairAccessory1/Preview/KittenEars1.png", 800, 290, 100, 100);
         DrawTextFit(
             "Ears",
             510,
@@ -2939,6 +3204,7 @@ CommandCombine([
             "Black"
         );
         DrawButton(500, getYPos(2), 400, 85, "", "White");
+        //DrawImageResize("Assets/Female3DCG/TailStraps/Preview/TailStrap.png", 800, 400, 80, 80);
         DrawTextFit(
             "Tails",
             510,
@@ -2947,6 +3213,7 @@ CommandCombine([
             "Black"
         );
         DrawButton(500, getYPos(3), 400, 85, "", "White");
+        //DrawImageResize("Assets/Female3DCG/Wings/AngelWings.png", 800, 490, 100, 100);
         DrawTextFit(
             "Wings",
             510,
@@ -2970,9 +3237,20 @@ CommandCombine([
             380,
             "Black"
         );
+        DrawButton(500, getYPos(6), 400, 85, "", "White");
+        DrawTextFit(
+            "Reactions",
+            510,
+            getYPos(6) + 42,
+            380,
+            "Black"
+        );
+        settings.draw("gender");
+        settings.draw("animal");
+        settings.draw("animationButtonsPosition");
+
         MainCanvas.textAlign = prev;
     };
-
     w.PreferenceSubscreenBCARSettingsExit = function() {
         PreferenceSubscreen = "";
         PreferenceSettings = "";
@@ -2991,6 +3269,22 @@ CommandCombine([
         if (MouseIn(500, getYPos(3), 400, 85)) LoadPreferencesSubscreen("Wings");
         if (MouseIn(500, getYPos(4), 400, 85)) LoadPreferencesSubscreen("Misc");
         if (MouseIn(500, getYPos(5), 400, 85)) LoadPreferencesSubscreen("Profiles");
+        if (MouseIn(500, getYPos(6), 400, 85)) LoadPreferencesSubscreen("Reactions");
+        if (settings.click("gender")){
+            let showMessage = true
+            const argsListGender = [settings.value("gender")];
+            CommandGenderToggle(argsListGender, false);
+        };
+        if (settings.click("animal")){
+            let showMessage = false
+            const argsListAnimals = [settings.value("animal")];
+            CommandAnimals(argsListAnimals, false);
+        };
+        if (settings.click("animationButtonsPosition")){
+            let showMessage = false
+            const argsListAnimationButtons = [settings.value("animationButtonsPosition")];
+            CommandAnimationButtons(argsListAnimationButtons, false);
+        };
 
         return;
     };
@@ -3107,6 +3401,7 @@ CommandCombine([
                     '/bcar cat - Changes the reactions and sounds to cat realted ones.',
                     '/bcar dog - Changes the reactions and sounds to dog realted ones.',
                     '/bcar fox - Changes the reactions and sounds to fox realted ones.',
+                    '/bcar human - Literally disables the reactions and sounds.',
                     '/bcar mouse - Changes the reactions and sounds to mouse realted ones.',
                 ],
             },
@@ -3197,6 +3492,8 @@ CommandCombine([
                     '/bcar wingdelay - Determines the flap speed.',
                     '/bcar wingdelete - Removes the wings.',
                     '/bcar winghelp - Opens wing instructions and commands page.',
+                    '/bcar fly - Starts flying.',
+                    '/bcar land - Stops flying.',
                 ],
             },
         ],